+++ lib/java/ydoc-server/src/main/java/org/enso/ydoc/server/Ydoc.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc;
+package org.enso.ydoc.server;
 
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
