--- lib/java/ydoc-server/src/test/java/org/enso/ydoc/jsonrpc/model/result/FileListResult.java
@@ -1,6 +1,0 @@
-package org.enso.ydoc.jsonrpc.model.result;
-
-import java.util.List;
-import org.enso.ydoc.jsonrpc.model.FileSystemObject;
-
-public record FileListResult(List<FileSystemObject> paths) implements Result {}
