--- lib/java/ydoc-server/src/test/java/org/enso/ydoc/jsonrpc/model/result/Result.java
@@ -1,3 +1,0 @@
-package org.enso.ydoc.jsonrpc.model.result;
-
-public interface Result {}
