+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/YdocTest.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc;
+package org.enso.ydoc.server;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -18,15 +18,15 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
-import org.enso.ydoc.jsonrpc.JsonRpcRequest;
-import org.enso.ydoc.jsonrpc.JsonRpcResponse;
-import org.enso.ydoc.jsonrpc.model.ContentRoot;
-import org.enso.ydoc.jsonrpc.model.FilePath;
-import org.enso.ydoc.jsonrpc.model.FileSystemObject;
-import org.enso.ydoc.jsonrpc.model.WriteCapability;
-import org.enso.ydoc.jsonrpc.model.result.FileListResult;
-import org.enso.ydoc.jsonrpc.model.result.InitProtocolConnectionResult;
-import org.enso.ydoc.jsonrpc.model.result.TextOpenFileResult;
+import org.enso.ydoc.server.jsonrpc.JsonRpcRequest;
+import org.enso.ydoc.server.jsonrpc.JsonRpcResponse;
+import org.enso.ydoc.server.jsonrpc.model.ContentRoot;
+import org.enso.ydoc.server.jsonrpc.model.FilePath;
+import org.enso.ydoc.server.jsonrpc.model.FileSystemObject;
+import org.enso.ydoc.server.jsonrpc.model.WriteCapability;
+import org.enso.ydoc.server.jsonrpc.model.result.FileListResult;
+import org.enso.ydoc.server.jsonrpc.model.result.InitProtocolConnectionResult;
+import org.enso.ydoc.server.jsonrpc.model.result.TextOpenFileResult;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
