+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/JsonRpcRequest.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.jsonrpc;
+package org.enso.ydoc.server.jsonrpc;
 
 import com.fasterxml.jackson.databind.JsonNode;
