+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/model/WriteCapability.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.jsonrpc.model;
+package org.enso.ydoc.server.jsonrpc.model;
 
 public record WriteCapability(String method, Options registerOptions) {
