+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/model/result/FileListResult.java
@@ -1,0 +1,6 @@
+package org.enso.ydoc.server.jsonrpc.model.result;
+
+import java.util.List;
+import org.enso.ydoc.server.jsonrpc.model.FileSystemObject;
+
+public record FileListResult(List<FileSystemObject> paths) implements Result {}
