+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/model/result/InitProtocolConnectionResult.java
@@ -1,7 +1,7 @@
-package org.enso.ydoc.jsonrpc.model.result;
+package org.enso.ydoc.server.jsonrpc.model.result;
 
 import java.util.List;
-import org.enso.ydoc.jsonrpc.model.ContentRoot;
+import org.enso.ydoc.server.jsonrpc.model.ContentRoot;
 
 public record InitProtocolConnectionResult(
     String ensoVersion, String currentEdition, List<ContentRoot> contentRoots) implements Result {}
