+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/model/result/Result.java
@@ -1,0 +1,3 @@
+package org.enso.ydoc.server.jsonrpc.model.result;
+
+public interface Result {}
