+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/server/jsonrpc/model/result/TextOpenFileResult.java
@@ -1,6 +1,6 @@
-package org.enso.ydoc.jsonrpc.model.result;
+package org.enso.ydoc.server.jsonrpc.model.result;
 
-import org.enso.ydoc.jsonrpc.model.WriteCapability;
+import org.enso.ydoc.server.jsonrpc.model.WriteCapability;
 
 public record TextOpenFileResult(
     WriteCapability writeCapability, String content, String currentVersion) implements Result {}
