+++ lib/java/ydoc/src/main/java/org/enso/ydoc/Polyfill.java
