+++ lib/java/ydoc/src/main/java/org/enso/ydoc/polyfill/web/Performance.java
