+++ lib/java/ydoc/src/main/resources/org/enso/ydoc/polyfill/web/performance.js
