+++ lib/java/ydoc/src/main/resources/org/enso/ydoc/polyfill/web/websocket.js
